Delete order items in order product lookup when the order is full refund
The previous commit will add the order item with negative net revenue
when the order is fully refunded. But the items sold in the products
analytics page is incorrect, because there was a DB record of an order item with
positive net revenue, and the product analytics query simply sums up all
the positive net revenue.