Fix duplicate spec evaluation in `evaluate_specs()` (#51166)
* feat: Add memoization to EvaluateSuggestion class

This commit adds memoization to the `EvaluateSuggestion` class in order to improve performance by caching the results of the `evaluate_specs` method. The memoization is implemented using an associative array called `$memo`, which stores the results of previous evaluations based on the input specifications and logger arguments. The memoization logic checks if the results for a given set of specifications and logger arguments already exist in the `$memo` array, and if so, returns the cached results instead of re-evaluating the specs. This helps to avoid redundant computations and improves the overall efficiency of the `evaluate_specs` method.

* Add changelog

* Only use specs as key

* Move shipping test to correct folder

* Fix tests

* Fix tests

* Fix lint

* Address PR feedback