Add the RuntimeContainer class.
This class is a replacement for ExtendedContainer as the underlying
dependency injection container used in WooCommerce at runtime.
A separate class will be developed as a replacement for unit test run time.

It's possible to force WooCommerce to sill use ExtendedContainer
at runtime by either creating a WOOCOMMERCE_USE_OLD_DI_CONTAINER
constat with a value of 'true', or hooking on the
'woocommerce_use_old_di_container' filter and returning true.