Add option to toggle which months method to use
Instead of always using the simplified query for the months filter,
which might trade worse UX for negligible performance gains for stores
with small, sporadic order volume, let's instead keep using the current
query method, and just add a toggle. This way the high volume stores can
enable the simple query as part of their tuning process, and other
stores won't see any change.

This also allows us to make use of the caching from the current method
when using the simplified method.

This also adds a unit test for the simplified method.