Alter styles for Woo emails when created via block editor
In the flow in WC_Email the content is generated and after styles are inlined.
We replace the email content when it is generated, but we need to provide styles
for inlining, which is done later in the flow.
We can apply only the block editor CSS or combine the editor's CSS with Woo email
CSS. I chose to combine because Woo email CSS contains some styles for the content which the editor's CSS doesn't cover.
I removed font-family and color styles from the Woo CSS to keep a unified look of the content.