Account for timezone offset from GMT
Ensure that the month options reflect the months in the site's timezone.
The order date is stored in GMT, so it could end up being a different
month and/or year than the local order date.

Also updates the months query so that the trash status is included in
the SQL statement via the prepare method, rather than interpolation.