Restore sender options and hide email design options
When the email editor is enabled, we don't want to display the email
design settings because the design is handled by the editor.
We also want to restore the sender's options to make it easy to edit,
because, as for now, we don't have direct access to the template edit mode.