Introduce isCustomerShippingDataUpdating and isCustomerBillingDataUpdating to handle separate loading states for checkout blocks
when we change any address we send to the server both billing and shipping addresses and isCustomerDateUpdating is used to detect the update.
With these changes the goal is to optimize loading states.

For example, for the shipping area we're only looking at isCustomerShippingDataUpdating.
isCalculating is not impacted, but isLoadingRates will only be true when only the shipping address changes.

Moving forward, we should optimize the shipping/billing updates only for the fields that would result in some checkout state changes.
For example, names and email don't impact the cart total.

I chose to keep the isCustomerDateUpdating because it's useful for blocking the Place order button, and also because it's used by 3PD.