Update Extensibility in Blocks documentation to resolve 404 errors
Originally reported on Slack at p1745429460671199-slack-C8X6Q7XQU

Many of the links on this page use a relative path (e.g., `../../plugins/woocommerce/client/blocks/`) to link out to Markdown files hosted in the GitHub repository.

When the page is published, the links use the relative path of `developer.woocommerce.com`, but since these pages don't exist, the resulting pages can't be found, generating a 404 error.

In this change, the links are updated to remove the relative path structure and they include the full GitHub Markdown file URLs. This will prevent 404 errors from occurring when navigating through this document.