Fix the useSelect warning about globalStylesPost
There was a strange fallback to template post when global styles post was not present.
After I removed the fallback I had to make a few fixes in style hooks.

Warning was:

The `useSelect` hook returns different values when called with the same state and parameters.
This can lead to unnecessary re-renders and performance issues if not fixed.

Non-equal value keys: globalStylesPost

WOOPLUG-3951