feat: unify filter data type, use array for all filter types
By using array for all filter types, we can have a simple check to
ensure we only set and use valid cache. We also have a guard for the
escape hatch as well, which enforce the filter data to be an array.

I thought about schema validation, but I don't want to hurt the
performance. It also sounds premature optimization.