fix: ensure the "empty" filter block is hidden after navigation
In other filter blocks except Price, we inject the hidden attribute
through the `get_block_wrapper_attributes()` function, which sets the
attribute like this: `hidden="1"`. iAPI keeps this attribute so it's
injected to the DOM after navigation.

The Price Filter block doesn't use the `get_block_wrapper_attributes()`
function, but declares the `hidden` attribute directly, without the
value (`="1"`). iAPI discards that attribute so the "empty" Price filter
block is visible after navigation.

This PR fixes the issue by using the `get_block_wrapper_attributes()`
function for the Price filter block as well. In addition, we add a
`.hidden` class and use it for all filter blocks to better control the
visibility of the filter blocks. The `hidden` attribute is still used
for accessibility reasons.