Turn the order into a draft when creating via REST API fails
If an exception is throw during the processing of the
POST /wc/v3/orders endpoint when the order has been already
created, the order is set to 'checkout-draft' state, and the
order id is returned inside 'data' in a 'new_draft_order_id' key.

Additionally, the code had a catch block for WC_REST_Exception
right after one for WC_Data_Exception, which is pointless because
the former inherits from the later and the second catch block
would never run (also the inherited class adds nothing to the
base class). Thus now only WC_Data_Exception is catched.