Fix the styles preview component to reflect changes
The main issue was that the Preview component was using static settings
value when reading style, which was set by the server.

I updated it to use live styles from the useEmailStyles hook, but I had to do additional
changes to be able to get values when the style was set as CSS variable.

WOOPLUG-4200