Fix that the theme color palette was not available in style > color settings
The color selection component uses theme settings to generate color palettes.
The site palette was added only to editor settings but was not present in the base theme,
which is used for the color selection component in the ScreenColors component.

This commit moves the logic for extracting handpicked site theme settings and styles
to a proper place.

WOOPLUG-4200