Remove the CurriedSelectorsOf from the @wordpress/data definition overrides
CurriedSelectorsOf is used to remove the `state` argument from the store selectors. However, we are currently using the types for selectors from `@types/wordpress__*` which already have the state argument removed. In this case, CurriedSelectorsOf would still remove the first argument, despite it's not the state, resulting in incorrect types for the selectors.