Remove hardcoded heading color default
This improves alignment with the site's global styles.
Heading color can be unset, and the heading will use the text color.
In case there is an extra color set for the headings, it is set
by WordPress's style system on the nested h1 element.

This also fixes an issue with Outlook, which was ignoring the color
set on the main wrapper.

WOOPLUG-4200