Refactor list view, open, close tracking, and inserter open tracking
This commit unifies this event with the rest of the events that track
isOpened within the event data instead of tracking separate events
for open and close.

It also moves the tracking to the DOM to track only interactions
via the buttons in the header. The panels are sometimes programmatically closed
for example, when a user selects distraction-free mode.

WOOPLUG-3953