Remove unit tests
Unit tests were too stubborn and kept failing as the test
environment works differently and, for some unknown
reason, doesn't throw an instance of `WP_Error`.

After many hours of debugging, the only solution I found
included changing the controller code to do manual SQL
queries. Doing such a big change for the sake of making
the unit test for this edge case work was not worth.

So, though it's a shame, I reverted and removed the unit
tests completely.