Refactor filters so that they are not used as a side effect
When a filter is used as a side effect, it sets a requirement that an integrator has to
hook into the filter before the editor JS file is loaded.

This requirement makes it harder for integrators to register their hooks.

This commit refactors all usages of applyFilter that they are no longer called
as side effects.
WOOPLUG-3653