Email Editor - prepare JS package (#58874)
* Update build to be aligned with other packages in the monorepo

WOOPLUG-3653

* Update the integration code in the Woo plugin to work with the updated build

WOOPLUG-3653

* Remove css imports from email editor compoents

This doesn't work after we aligned Woo packages builds.
We no longer have webpack plugin for that.
We build css directly from styles.scss.
WOOPLUG-3653

* Extend the common tsconfig.json in the email editor tsconfig.json

I kept our config for rules and I chose to use
"jsx": "react-jsx" which is more modern option, which can be used
with React 17+. We don't need to import react to every component
with "jsx": "react-jsx" it is added automatically.
WOOPLUG-3653

* Reintroduce prettier to the email editor

It was removed in the big update of configs in the first commit.
WOOPLUG-3653

* Export initializeEditor function instead of starting the editor as a sideeffect

WOOPLUG-3653

* Reintroduce stylelint

WOOPLUG-3653

* Sync dependencies in the monorepo

WOOPLUG-3653

* Update the editor package's readme.md

WOOPLUG-3653

* Rename composer project in JS package

WOOPLUG-3653

* Refactor filters so that they are not used as a side effect

When a filter is used as a side effect, it sets a requirement that an integrator has to
hook into the filter before the editor JS file is loaded.

This requirement makes it harder for integrators to register their hooks.

This commit refactors all usages of applyFilter that they are no longer called
as side effects.
WOOPLUG-3653

* Add changelogs

WOOPLUG-3653

* Skip copying rich-text.js asset

We don't modify it so it makes no sense to copy as we can include it directly.
We will remove this asset when we stop supporting WP 6.7
WOOPLUG-3653

* Cleanup devDependencies

These depedencies are provided by @woocommerce/internal-style-build
WOOPLUG-3653

* Monorepo: adjustments to email-editor build.

* Install email-editor deps after rebase

WOOPLUG-3653

---------

Co-authored-by: Vladimir Reznichenko <kalessil@gmail.com>