Optimize switch_blog hook
`switch_blog` is triggered in a multisite install when iterating over
many sites. With large network this will result in the hook being triggerd
hundreds or thousands of times during a single request, this results in the
`$wpdb->tables` array being filled with duplicate table names. This optimization
only modifies prefixed table and stops filling the array with duplicates.