Add `role="alert"` to `woocommerce-info` type notices.
In #44283, when moving new notice styles to display in block themes only,
there was an oversight in removing the `role="alert"` for the classic info notices.

The other notices, `success` and `error`, already have `role="alert"` applied to
them, so that was inconsistent and meant that `info` notices were not read by
screen readers. This affected all extensions that used `wc_add_notice` to display
info notices.