Move Add note for orders to the top of the note box.
The notes display with the newest first, so when adding a note you would
usually be looking a the newest note which is also at the top.
This locates the add note box close to the most recent note.
Otherwise if you have a long list of notes you would be looking at the most recent note at the top and then have to scroll all the way to the bottom to add the new note.

Fixes #34648