Improve the block checking in ProductCollection/Utils by using has_block
It's not performant to do a full parsing of all blocks to detect the
presence of a block. There might be some situation where you need
more accuracy such as in detecting block attributes, but this method
is not used anywhere else, and has questionable benefit as a public
API since one can parse the block tree themselves quite easily.

has_block is a simpler string search and should perform better. To
avoid tripping others with using this existing block detection util
I've deprecated it.