Fix calculate_cogs_value_core for product order items.
PR #61307 introduced a change: already persisted COGS values for
product line items would be preserved to serve as "snapshots"
of the COGS value at the time of purchase. However this has the
unintended side effect that for orders not yet completed, the COGS
values for these line items will not be recalculated even after
product COGS change. So this commit reverts the change.