Pin Old `coenjacobs/mozart` Version
Unfortunately, `dev-master` will no longer work because
the PHP requirement was bumped to 8.0. In order to
make sure we can still use PHP 7.3 binaries in tests,
I have downgraded the package. This should be
fine as it is still older than the 0.6.0 that we
used when we integrated the package.