Add 'skip_initial_save' option to settings and update installation logic
When set to true, this allows the plugin to skip saving the initial option value in the db for the settings field.

In this case, we don't want to save the initial value for the woocommerce_email_from_name field. This will allow us to continue using the blog title for the email “from” until the user manually updates it or saves the email settings field, which will repopulate the option.