Allow Unbounded Term Requests
Since the default is `0` and most requests would have been
written with that expectation, the silent backward
compatibility break isn't worth the improvement.

In practice, term endpoints aren't going to consume much memory.
There is a strong case for always fetching all of the terms, such as
to display options in a user interface or cache them somewhere.